+++ .gitignore
@@ -155,7 +155,6 @@ integration_tests/scripts/main.tf
 integration_tests/scripts/.terraform
 .DS_Store
 snmp-ssh-key.pem
-new_integration_testing
 
 # Sonarqube
 .scannerwork
