+++ integration_tests/test_poller_integration.py
@@ -1,55 +1,39 @@
+#   ########################################################################
+#   Copyright 2021 Splunk Inc.
+#
+#    Licensed under the Apache License, Version 2.0 (the "License");
+#    you may not use this file except in compliance with the License.
+#    You may obtain a copy of the License at
+#
+#        http://www.apache.org/licenses/LICENSE-2.0
+#
+#    Unless required by applicable law or agreed to in writing, software
+#    distributed under the License is distributed on an "AS IS" BASIS,
+#    WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+#    See the License for the specific language governing permissions and
+#    limitations under the License.
+#   ########################################################################
 import logging
-import subprocess
 import time
 
 import pytest
 from ruamel.yaml.scalarstring import DoubleQuotedScalarString as dq
 from ruamel.yaml.scalarstring import SingleQuotedScalarString as sq
 
-from integration_tests.compose_deployment.splunk_test_utils_compose import (
+from integration_tests.splunk_test_utils import (
     splunk_single_search,
-    update_groups,
-    update_inventory,
-    update_profiles,
+    update_file_microk8s,
+    update_groups_compose,
+    update_groups_microk8s,
+    update_inventory_compose,
+    update_profiles_compose,
+    update_profiles_microk8s,
     upgrade_docker_compose,
+    upgrade_helm_microk8s,
     yaml_escape_list,
 )
 
 logger = logging.getLogger(__name__)
-
-
-def logs_from_container(container):
-    result = subprocess.run(
-        ["sudo", "docker", "logs", f"{container}"],
-        stdout=subprocess.PIPE,  # Capture standard output
-        stderr=subprocess.PIPE,  # Capture standard error (optional)
-        text=True,  # Ensure the output is returned as a string
-    )
-    output = result.stdout
-    error_output = result.stderr
-    logger.info(f"{container} logs:\n{output}")
-    if error_output:
-        logger.error(f"{container} logs errors:\n{error_output}")
-
-
-def docker_ps():
-    result = subprocess.run(
-        ["sudo", "docker", "ps"],
-        stdout=subprocess.PIPE,  # Capture standard output
-        stderr=subprocess.PIPE,  # Capture standard error (optional)
-        text=True,  # Ensure the output is returned as a string
-    )
-    output = result.stdout
-    error_output = result.stderr
-    logger.info(f"docker ps:\n{output}")
-    if error_output:
-        logger.error(f"docker ps errors:\n{error_output}")
-
-
-def print_file(filename):
-    with open(filename) as f:
-        lines = f.readlines()
-        logger.info(f"file {filename}:\n{lines}")
 
 
 class TestSanity:
@@ -58,8 +42,8 @@ class TestSanity:
         search_string = (
             """search index="netops" sourcetype="sc4snmp:event" earliest=-5m"""
         )
-
         result_count, events_count = splunk_single_search(setup_splunk, search_string)
+
         assert result_count > 0
         assert events_count > 0
 
@@ -93,6 +77,7 @@ class TestSanity:
 @pytest.fixture(scope="class")
 def setup_profile(request):
     trap_external_ip = request.config.getoption("trap_external_ip")
+    deployment = request.config.getoption("sc4snmp_deployment")
     profile = {
         "generic_switch": {
             "frequency": 5,
@@ -102,13 +87,30 @@ def setup_profile(request):
             ],
         }
     }
-    update_profiles(profile)
-    update_inventory([f"{trap_external_ip},,2c,public,,,600,generic_switch,,"])
-    upgrade_docker_compose()
+
+    if str(deployment) == "microk8s":
+        update_profiles_microk8s(profile)
+        update_file_microk8s(
+            [f"{trap_external_ip},,2c,public,,,600,generic_switch,,"], "inventory.yaml"
+        )
+        upgrade_helm_microk8s(["inventory.yaml", "profiles.yaml"])
+    else:
+        update_profiles_compose(profile)
+        update_inventory_compose(
+            [f"{trap_external_ip},,2c,public,,,600,generic_switch,,"]
+        )
+        upgrade_docker_compose()
     time.sleep(30)
     yield
-    update_inventory([f"{trap_external_ip},,2c,public,,,600,generic_switch,,t"])
-    upgrade_docker_compose()
+    if str(deployment) == "microk8s":
+        upgrade_helm_microk8s(
+            [f"{trap_external_ip},,2c,public,,,600,generic_switch,,t"]
+        )
+    else:
+        update_inventory_compose(
+            [f"{trap_external_ip},,2c,public,,,600,generic_switch,,t"]
+        )
+        upgrade_docker_compose()
     time.sleep(20)
 
 
@@ -136,6 +138,7 @@ class TestProfiles:
 @pytest.fixture(scope="class")
 def setup_profiles(request):
     trap_external_ip = request.config.getoption("trap_external_ip")
+    deployment = request.config.getoption("sc4snmp_deployment")
     profile = {
         "new_profile": {"frequency": 7, "varBinds": [yaml_escape_list(sq("IP-MIB"))]},
         "generic_switch": {
@@ -143,17 +146,32 @@ def setup_profiles(request):
             "varBinds": [yaml_escape_list(sq("UDP-MIB"))],
         },
     }
-    update_profiles(profile)
-    update_inventory(
-        [f"{trap_external_ip},,2c,public,,,600,new_profile;generic_switch,,"]
-    )
-    upgrade_docker_compose()
+    if deployment == "microk8s":
+        update_profiles_microk8s(profile)
+        update_file_microk8s(
+            [f"{trap_external_ip},,2c,public,,,600,new_profile;generic_switch,,"],
+            "inventory.yaml",
+        )
+        upgrade_helm_microk8s(["inventory.yaml", "profiles.yaml"])
+    else:
+        update_profiles_compose(profile)
+        update_inventory_compose(
+            [f"{trap_external_ip},,2c,public,,,600,new_profile;generic_switch,,"]
+        )
+        upgrade_docker_compose()
     time.sleep(30)
     yield
-    update_inventory(
-        [f"{trap_external_ip},,2c,public,,,600,new_profile;generic_switch,,t"]
-    )
-    upgrade_docker_compose()
+    if deployment == "microk8s":
+        update_file_microk8s(
+            [f"{trap_external_ip},,2c,public,,,600,new_profile;generic_switch,,t"],
+            "inventory.yaml",
+        )
+        upgrade_helm_microk8s(["inventory.yaml"])
+    else:
+        update_inventory_compose(
+            [f"{trap_external_ip},,2c,public,,,600,new_profile;generic_switch,,t"]
+        )
+        upgrade_docker_compose()
     time.sleep(20)
 
 
@@ -161,16 +179,15 @@ def setup_profiles(request):
 class TestProfilesWorkflow:
     def test_add_new_profile_and_reload(self, setup_splunk):
         search_string = """| mpreview index=netmetrics| spath profiles | search profiles=new_profile """
-
         result_count, metric_count = run_retried_single_search(
             setup_splunk, search_string, 2
         )
-
         assert result_count > 0
         assert metric_count > 0
 
     def test_disable_one_profile_and_reload(self, request, setup_splunk):
         trap_external_ip = request.config.getoption("trap_external_ip")
+        deployment = request.config.getoption("sc4snmp_deployment")
         logger.info("Integration test for deleting one profile and reloading")
         profile = {
             "new_profile": {
@@ -178,31 +195,46 @@ class TestProfilesWorkflow:
                 "varBinds": [yaml_escape_list(sq("IP-MIB"))],
             }
         }
-        update_profiles(profile)
-        update_inventory([f"{trap_external_ip},,2c,public,,,600,new_profile,,"])
-        upgrade_docker_compose()
+        if deployment == "microk8s":
+            update_profiles_microk8s(profile)
+            update_file_microk8s(
+                [f"{trap_external_ip},,2c,public,,,600,new_profile,,"], "inventory.yaml"
+            )
+            upgrade_helm_microk8s(["inventory.yaml", "profiles.yaml"])
+        else:
+            update_profiles_compose(profile)
+            update_inventory_compose(
+                [f"{trap_external_ip},,2c,public,,,600,new_profile,,"]
+            )
+            upgrade_docker_compose()
         time.sleep(70)
-
         search_string = """| mpreview index=netmetrics| spath profiles | search profiles=generic_switch earliest=-20s """
         result_count, metric_count = run_retried_single_search(
             setup_splunk, search_string, 2
         )
-
         assert result_count == 0
         assert metric_count == 0
 
     def test_delete_inventory_line(self, request, setup_splunk):
         trap_external_ip = request.config.getoption("trap_external_ip")
+        deployment = request.config.getoption("sc4snmp_deployment")
         logger.info("Integration test for deleting one profile and reloading")
-        update_inventory([f"{trap_external_ip},,2c,public,,,600,new_profile,,t"])
-        upgrade_docker_compose()
+        if deployment == "microk8s":
+            update_file_microk8s(
+                [f"{trap_external_ip},,2c,public,,,600,new_profile,,t"],
+                "inventory.yaml",
+            )
+            upgrade_helm_microk8s(["inventory.yaml", "profiles.yaml"])
+        else:
+            update_inventory_compose(
+                [f"{trap_external_ip},,2c,public,,,600,new_profile,,t"]
+            )
+            upgrade_docker_compose()
         time.sleep(40)
-
         search_string = """| mpreview index=netmetrics earliest=-20s """
         result_count, metric_count = run_retried_single_search(
             setup_splunk, search_string, 2
         )
-
         assert result_count == 0
         assert metric_count == 0
 
@@ -210,6 +242,7 @@ class TestProfilesWorkflow:
 @pytest.fixture(scope="class")
 def setup_smart_profiles(request):
     trap_external_ip = request.config.getoption("trap_external_ip")
+    deployment = request.config.getoption("sc4snmp_deployment")
     logger.info("Integration test for fields smart profiles")
     profile = {
         "smart_profile_field": {
@@ -225,13 +258,26 @@ def setup_smart_profiles(request):
             ],
         }
     }
-    update_profiles(profile)
-    update_inventory([f"{trap_external_ip},,2c,public,,,600,,t,"])
-    upgrade_docker_compose()
+    if deployment == "microk8s":
+        update_profiles_microk8s(profile)
+        update_file_microk8s(
+            [f"{trap_external_ip},,2c,public,,,600,,t,"], "inventory.yaml"
+        )
+        upgrade_helm_microk8s(["inventory.yaml", "profiles.yaml"])
+    else:
+        update_profiles_compose(profile)
+        update_inventory_compose([f"{trap_external_ip},,2c,public,,,600,,t,"])
+        upgrade_docker_compose()
     time.sleep(30)
     yield
-    update_inventory([f"{trap_external_ip},,2c,public,,,600,,t,t"])
-    upgrade_docker_compose()
+    if deployment == "microk8s":
+        update_file_microk8s(
+            [f"{trap_external_ip},,2c,public,,,600,,t,t"], "inventory.yaml"
+        )
+        upgrade_helm_microk8s(["inventory.yaml"])
+    else:
+        update_inventory_compose([f"{trap_external_ip},,2c,public,,,600,,t,t"])
+        upgrade_docker_compose()
     time.sleep(20)
 
 
@@ -258,11 +304,11 @@ class TestSmartProfiles:
 
     def test_smart_profiles_base(self, setup_splunk):
         logger.info("Integration test for fields base smart profiles")
+        time.sleep(300)
         search_string_baseIF = (
             """| mpreview index=netmetrics| spath profiles | search profiles=BaseIF """
         )
         search_string_baseUpTime = """| mpreview index=netmetrics| spath profiles | search profiles=BaseUpTime """
-        time.sleep(300)
         result_count, metric_count = splunk_single_search(
             setup_splunk, search_string_baseIF
         )
@@ -278,19 +324,37 @@ class TestSmartProfiles:
 @pytest.fixture
 def setup_modify_profile(request):
     trap_external_ip = request.config.getoption("trap_external_ip")
+    deployment = request.config.getoption("sc4snmp_deployment")
     profile = {
         "test_modify": {
             "frequency": 5,
             "varBinds": [yaml_escape_list(sq("UDP-MIB"))],
         },
     }
-    update_profiles(profile)
-    update_inventory([f"{trap_external_ip},,2c,public,,,600,test_modify,f,"])
-    upgrade_docker_compose()
+    if deployment == "microk8s":
+        update_profiles_microk8s(profile)
+        update_file_microk8s(
+            [f"{trap_external_ip},,2c,public,,,600,test_modify,f,"], "inventory.yaml"
+        )
+        upgrade_helm_microk8s(["inventory.yaml", "profiles.yaml"])
+    else:
+        update_profiles_compose(profile)
+        update_inventory_compose(
+            [f"{trap_external_ip},,2c,public,,,600,test_modify,f,"]
+        )
+        upgrade_docker_compose()
     time.sleep(30)
     yield
-    update_inventory([f"{trap_external_ip},,2c,public,,,600,test_modify,f,t"])
-    upgrade_docker_compose()
+    if deployment == "microk8s":
+        update_file_microk8s(
+            [f"{trap_external_ip},,2c,public,,,600,test_modify,f,t"], "inventory.yaml"
+        )
+        upgrade_helm_microk8s(["inventory.yaml"])
+    else:
+        update_inventory_compose(
+            [f"{trap_external_ip},,2c,public,,,600,test_modify,f,t"]
+        )
+        upgrade_docker_compose()
     time.sleep(20)
 
 
@@ -306,18 +370,41 @@ class TestModifyProfilesFrequency:
 
     def test_modify_frequency_field(self, request, setup_splunk):
         trap_external_ip = request.config.getoption("trap_external_ip")
+        deployment = request.config.getoption("sc4snmp_deployment")
         profile = {
             "test_modify": {
                 "frequency": 7,
                 "varBinds": [yaml_escape_list(sq("UDP-MIB"))],
             },
         }
-        update_profiles(profile)
-        update_inventory([f"{trap_external_ip},,2c,public,,,600,test_modify,f,t"])
-        upgrade_docker_compose()
+        if deployment == "microk8s":
+            update_profiles_microk8s(profile)
+            update_file_microk8s(
+                [f"{trap_external_ip},,2c,public,,,600,test_modify,f,t"],
+                "inventory.yaml",
+            )
+            upgrade_helm_microk8s(["inventory.yaml", "profiles.yaml"])
+        else:
+            update_profiles_compose(profile)
+            update_inventory_compose(
+                [f"{trap_external_ip},,2c,public,,,600,test_modify,f,t"]
+            )
+            upgrade_docker_compose()
+
         time.sleep(60)
-        update_inventory([f"{trap_external_ip},,2c,public,,,600,test_modify,f,"])
-        upgrade_docker_compose()
+
+        if deployment == "microk8s":
+            update_file_microk8s(
+                [f"{trap_external_ip},,2c,public,,,600,test_modify,f,"],
+                "inventory.yaml",
+            )
+            upgrade_helm_microk8s(["inventory.yaml", "profiles.yaml"])
+        else:
+            update_inventory_compose(
+                [f"{trap_external_ip},,2c,public,,,600,test_modify,f,"]
+            )
+            upgrade_docker_compose()
+
         time.sleep(30)
         search_string = """| mpreview index=netmetrics earliest=-30s | search profiles=test_modify frequency=7 """
         result_count, metric_count = run_retried_single_search(
@@ -339,6 +426,7 @@ class TestModifyProfilesVarBinds:
 
     def test_modify_varBinds_field(self, request, setup_splunk):
         trap_external_ip = request.config.getoption("trap_external_ip")
+        deployment = request.config.getoption("sc4snmp_deployment")
         profile = {
             "test_modify": {
                 "frequency": 7,
@@ -349,12 +437,34 @@ class TestModifyProfilesVarBinds:
                 ],
             },
         }
-        update_profiles(profile)
-        update_inventory([f"{trap_external_ip},,2c,public,,,600,test_modify,f,t"])
-        upgrade_docker_compose()
+        if deployment == "microk8s":
+            update_profiles_microk8s(profile)
+            update_file_microk8s(
+                [f"{trap_external_ip},,2c,public,,,600,test_modify,f,t"],
+                "inventory.yaml",
+            )
+            upgrade_helm_microk8s(["inventory.yaml", "profiles.yaml"])
+        else:
+            update_profiles_compose(profile)
+            update_inventory_compose(
+                [f"{trap_external_ip},,2c,public,,,600,test_modify,f,t"]
+            )
+            upgrade_docker_compose()
+
         time.sleep(60)
-        update_inventory([f"{trap_external_ip},,2c,public,,,600,test_modify,f,"])
-        upgrade_docker_compose()
+
+        if deployment == "microk8s":
+            update_file_microk8s(
+                [f"{trap_external_ip},,2c,public,,,600,test_modify,f,"],
+                "inventory.yaml",
+            )
+            upgrade_helm_microk8s(["inventory.yaml", "profiles.yaml"])
+        else:
+            update_inventory_compose(
+                [f"{trap_external_ip},,2c,public,,,600,test_modify,f,"]
+            )
+            upgrade_docker_compose()
+
         time.sleep(20)
         search_string = """| mpreview index=netmetrics earliest=-15s | search profiles=test_modify TCP-MIB """
         result_count, metric_count = run_retried_single_search(
@@ -379,19 +489,37 @@ class TestModifyProfilesVarBinds:
 @pytest.fixture
 def setup_small_walk(request):
     trap_external_ip = request.config.getoption("trap_external_ip")
+    deployment = request.config.getoption("sc4snmp_deployment")
     profile = {
         "walk1": {
             "condition": {"type": "walk"},
             "varBinds": [yaml_escape_list(sq("IP-MIB"))],
         },
     }
-    update_profiles(profile)
-    update_inventory([f"{trap_external_ip},,2c,public,,,20,walk1,f,"])
-    upgrade_docker_compose()
+
+    if deployment == "microk8s":
+        update_profiles_microk8s(profile)
+        update_file_microk8s(
+            [f"{trap_external_ip},,2c,public,,,20,walk1,f,"], "inventory.yaml"
+        )
+        upgrade_helm_microk8s(["inventory.yaml", "profiles.yaml"])
+    else:
+        update_profiles_compose(profile)
+        update_inventory_compose([f"{trap_external_ip},,2c,public,,,20,walk1,f,"])
+        upgrade_docker_compose()
+
     time.sleep(30)
+
     yield
-    update_inventory([f"{trap_external_ip},,2c,public,,,20,walk1,f,t"])
-    upgrade_docker_compose()
+    if deployment == "microk8s":
+        update_file_microk8s(
+            [f"{trap_external_ip},,2c,public,,,20,walk1,f,t"], "inventory.yaml"
+        )
+        upgrade_helm_microk8s(["inventory.yaml"])
+    else:
+        update_inventory_compose([f"{trap_external_ip},,2c,public,,,20,walk1,f,t"])
+        upgrade_docker_compose()
+
     time.sleep(20)
 
 
@@ -420,24 +548,42 @@ class TestSmallWalk:
 @pytest.fixture()
 def setup_v3_connection(request):
     trap_external_ip = request.config.getoption("trap_external_ip")
+    deployment = request.config.getoption("sc4snmp_deployment")
     time.sleep(60)
-    profile = {
-        "v3profile": {
-            "frequency": 5,
-            "varBinds": [
-                yaml_escape_list(sq("IF-MIB")),
-                yaml_escape_list(sq("TCP-MIB")),
-                yaml_escape_list(sq("UDP-MIB")),
-            ],
-        },
-    }
-    update_profiles(profile)
-    update_inventory([f"{trap_external_ip},1161,3,,sv3poller,,20,v3profile,f,"])
-    upgrade_docker_compose()
+    if deployment == "microk8s":
+        update_file_microk8s(
+            [f"{trap_external_ip},1161,3,,sv3poller,,20,v3profile,f,"], "inventory.yaml"
+        )
+        upgrade_helm_microk8s(["inventory.yaml"])
+    else:
+        profile = {
+            "v3profile": {
+                "frequency": 5,
+                "varBinds": [
+                    yaml_escape_list(sq("IF-MIB")),
+                    yaml_escape_list(sq("TCP-MIB")),
+                    yaml_escape_list(sq("UDP-MIB")),
+                ],
+            },
+        }
+        update_profiles_compose(profile)
+        update_inventory_compose(
+            [f"{trap_external_ip},1161,3,,sv3poller,,20,v3profile,f,"]
+        )
+        upgrade_docker_compose()
     time.sleep(30)
     yield
-    update_inventory([f"{trap_external_ip},1161,3,,sv3poller,,20,v3profile,f,t"])
-    upgrade_docker_compose()
+    if deployment == "microk8s":
+        update_file_microk8s(
+            [f"{trap_external_ip},1161,3,,sv3poller,,20,v3profile,f,t"],
+            "inventory.yaml",
+        )
+        upgrade_helm_microk8s(["inventory.yaml"])
+    else:
+        update_inventory_compose(
+            [f"{trap_external_ip},1161,3,,sv3poller,,20,v3profile,f,t"]
+        )
+        upgrade_docker_compose()
     time.sleep(20)
 
 
@@ -456,6 +602,7 @@ class TestSNMPv3Connection:
 @pytest.fixture(scope="class")
 def setup_groups(request):
     trap_external_ip = request.config.getoption("trap_external_ip")
+    deployment = request.config.getoption("sc4snmp_deployment")
     profiles = {
         "single_profile": {
             "frequency": 7,
@@ -481,26 +628,51 @@ def setup_groups(request):
         ],
     }
 
-    update_profiles(profiles)
-    update_groups(groups)
-    update_inventory(
-        [
-            f"{trap_external_ip},1165,2c,public,,,600,single_profile,,",
-            f"routers,,2c,public,,,600,routers_profile,,",
-            f"switches,,2c,public,,,600,switches_profile,,",
-        ]
-    )
-    upgrade_docker_compose()
+    if deployment == "microk8s":
+        update_profiles_microk8s(profiles)
+        update_groups_microk8s(groups)
+        update_file_microk8s(
+            [
+                f"{trap_external_ip},1165,2c,public,,,600,single_profile,,",
+                f"routers,,2c,public,,,600,routers_profile,,",
+                f"switches,,2c,public,,,600,switches_profile,,",
+            ],
+            "inventory.yaml",
+        )
+        upgrade_helm_microk8s(["inventory.yaml", "profiles.yaml", "groups.yaml"])
+    else:
+        update_profiles_compose(profiles)
+        update_groups_compose(groups)
+        update_inventory_compose(
+            [
+                f"{trap_external_ip},1165,2c,public,,,600,single_profile,,",
+                f"routers,,2c,public,,,600,routers_profile,,",
+                f"switches,,2c,public,,,600,switches_profile,,",
+            ]
+        )
+        upgrade_docker_compose()
+
     time.sleep(120)
     yield
-    update_inventory(
-        [
-            f"{trap_external_ip},1165,2c,public,,,600,single_profile,,t",
-            f"routers,,2c,public,,,600,routers_profile,,t",
-            f"switches,,2c,public,,,600,switches_profile,,t",
-        ]
-    )
-    upgrade_docker_compose()
+    if deployment == "microk8s":
+        update_file_microk8s(
+            [
+                f"{trap_external_ip},1165,2c,public,,,600,single_profile,,t",
+                f"routers,,2c,public,,,600,routers_profile,,t",
+                f"switches,,2c,public,,,600,switches_profile,,t",
+            ],
+            "inventory.yaml",
+        )
+        upgrade_helm_microk8s(["inventory.yaml"])
+    else:
+        update_inventory_compose(
+            [
+                f"{trap_external_ip},1165,2c,public,,,600,single_profile,,t",
+                f"routers,,2c,public,,,600,routers_profile,,t",
+                f"switches,,2c,public,,,600,switches_profile,,t",
+            ]
+        )
+        upgrade_docker_compose()
     time.sleep(100)
 
 
@@ -536,6 +708,7 @@ class TestGroupsInventory:
 
     def test_edit_routers_group(self, request, setup_splunk):
         trap_external_ip = request.config.getoption("trap_external_ip")
+        deployment = request.config.getoption("sc4snmp_deployment")
         new_groups = {
             "routers": [{"address": trap_external_ip, "port": 1164}],
             "switches": [
@@ -543,8 +716,13 @@ class TestGroupsInventory:
                 {"address": trap_external_ip, "port": 1162},
             ],
         }
-        update_groups(new_groups)
-        upgrade_docker_compose()
+        if deployment == "microk8s":
+            update_groups_microk8s(new_groups)
+            upgrade_helm_microk8s(["inventory.yaml", "profiles.yaml", "groups.yaml"])
+        else:
+            update_groups_compose(new_groups)
+            upgrade_docker_compose()
+
         time.sleep(60)
         search_string = f"""| mpreview index=netmetrics earliest=-20s | search profiles=routers_profile host="{trap_external_ip}:1163" """
         result_count, metric_count = run_retried_single_search(
@@ -557,6 +735,7 @@ class TestGroupsInventory:
 @pytest.fixture(scope="class")
 def setup_single_ang_group(request):
     trap_external_ip = request.config.getoption("trap_external_ip")
+    deployment = request.config.getoption("sc4snmp_deployment")
     profiles = {
         "single_profile_1": {
             "frequency": 7,
@@ -577,26 +756,52 @@ def setup_single_ang_group(request):
         ],
     }
 
-    update_profiles(profiles)
-    update_groups(groups)
-    update_inventory(
-        [
-            f"{trap_external_ip},1165,2c,public,,,600,single_profile_1,,",
-            f"{trap_external_ip},1162,2c,public,,,600,single_profile_2,,",
-            f"switches,,2c,public,,,600,switches_profile,,",
-        ]
-    )
-    upgrade_docker_compose()
+    if deployment == "microk8s":
+        update_profiles_microk8s(profiles)
+        update_groups_microk8s(groups)
+        update_file_microk8s(
+            [
+                f"{trap_external_ip},1165,2c,public,,,600,single_profile_1,,",
+                f"{trap_external_ip},1162,2c,public,,,600,single_profile_2,,",
+                f"switches,,2c,public,,,600,switches_profile,,",
+            ],
+            "inventory.yaml",
+        )
+        upgrade_helm_microk8s(["inventory.yaml", "profiles.yaml", "groups.yaml"])
+    else:
+        update_profiles_compose(profiles)
+        update_groups_compose(groups)
+        update_inventory_compose(
+            [
+                f"{trap_external_ip},1165,2c,public,,,600,single_profile_1,,",
+                f"{trap_external_ip},1162,2c,public,,,600,single_profile_2,,",
+                f"switches,,2c,public,,,600,switches_profile,,",
+            ]
+        )
+        upgrade_docker_compose()
+
     time.sleep(120)
     yield
-    update_inventory(
-        [
-            f"{trap_external_ip},1165,2c,public,,,600,single_profile_1,,t",
-            f"{trap_external_ip},1162,2c,public,,,600,single_profile_2,,t",
-            f"switches,,2c,public,,,600,switches_profile,,t",
-        ]
-    )
-    upgrade_docker_compose()
+
+    if deployment == "microk8s":
+        update_file_microk8s(
+            [
+                f"{trap_external_ip},1165,2c,public,,,600,single_profile_1,,t",
+                f"{trap_external_ip},1162,2c,public,,,600,single_profile_2,,t",
+                f"switches,,2c,public,,,600,switches_profile,,t",
+            ],
+            "inventory.yaml",
+        )
+        upgrade_helm_microk8s(["inventory.yaml"])
+    else:
+        update_inventory_compose(
+            [
+                f"{trap_external_ip},1165,2c,public,,,600,single_profile_1,,t",
+                f"{trap_external_ip},1162,2c,public,,,600,single_profile_2,,t",
+                f"switches,,2c,public,,,600,switches_profile,,t",
+            ]
+        )
+        upgrade_docker_compose()
     time.sleep(100)
 
 
@@ -644,6 +849,7 @@ def setup_single_gt_and_lt_profiles(request):
     lt_profile should result in polling IF-MIB.ifOutDiscards.2
     """
     trap_external_ip = request.config.getoption("trap_external_ip")
+    deployment = request.config.getoption("sc4snmp_deployment")
     profiles = {
         "gt_profile": {
             "frequency": 7,
@@ -661,17 +867,36 @@ def setup_single_gt_and_lt_profiles(request):
         },
     }
 
-    update_profiles(profiles)
-    update_inventory(
-        [f"{trap_external_ip},1166,2c,public,,,600,gt_profile;lt_profile,,"]
-    )
-    upgrade_docker_compose()
+    if deployment == "microk8s":
+        update_profiles_microk8s(profiles)
+        update_file_microk8s(
+            [
+                f"{trap_external_ip},1166,2c,public,,,600,gt_profile;lt_profile,,",
+            ],
+            "inventory.yaml",
+        )
+        upgrade_helm_microk8s(["inventory.yaml", "profiles.yaml"])
+    else:
+        update_profiles_compose(profiles)
+        update_inventory_compose(
+            [f"{trap_external_ip},1166,2c,public,,,600,gt_profile;lt_profile,,"]
+        )
+        upgrade_docker_compose()
     time.sleep(120)
     yield
-    update_inventory(
-        [f"{trap_external_ip},1166,2c,public,,,600,gt_profile;lt_profile,,t"]
-    )
-    upgrade_docker_compose()
+    if deployment == "microk8s":
+        update_file_microk8s(
+            [
+                f"{trap_external_ip},1166,2c,public,,,600,gt_profile;lt_profile,,t",
+            ],
+            "inventory.yaml",
+        )
+        upgrade_helm_microk8s(["inventory.yaml"])
+    else:
+        update_inventory_compose(
+            [f"{trap_external_ip},1166,2c,public,,,600,gt_profile;lt_profile,,t"]
+        )
+        upgrade_docker_compose()
     time.sleep(120)
 
 
@@ -707,6 +932,7 @@ def setup_single_in_and_equals_profiles(request):
     equals_profile should result in polling IF-MIB.ifOutDiscards.1
     """
     trap_external_ip = request.config.getoption("trap_external_ip")
+    deployment = request.config.getoption("sc4snmp_deployment")
     profiles = {
         "in_profile": {
             "frequency": 7,
@@ -732,17 +958,36 @@ def setup_single_in_and_equals_profiles(request):
         },
     }
 
-    update_profiles(profiles)
-    update_inventory(
-        [f"{trap_external_ip},1166,2c,public,,,600,in_profile;equals_profile,,"]
-    )
-    upgrade_docker_compose()
+    if deployment == "microk8s":
+        update_profiles_microk8s(profiles)
+        update_file_microk8s(
+            [
+                f"{trap_external_ip},1166,2c,public,,,600,in_profile;equals_profile,,",
+            ],
+            "inventory.yaml",
+        )
+        upgrade_helm_microk8s(["inventory.yaml", "profiles.yaml"])
+    else:
+        update_profiles_compose(profiles)
+        update_inventory_compose(
+            [f"{trap_external_ip},1166,2c,public,,,600,in_profile;equals_profile,,"]
+        )
+        upgrade_docker_compose()
     time.sleep(120)
     yield
-    update_inventory(
-        [f"{trap_external_ip},1166,2c,public,,,600,in_profile;equals_profile,,t"]
-    )
-    upgrade_docker_compose()
+    if deployment == "microk8s":
+        update_file_microk8s(
+            [
+                f"{trap_external_ip},1166,2c,public,,,600,in_profile;equals_profile,,t",
+            ],
+            "inventory.yaml",
+        )
+        upgrade_helm_microk8s(["inventory.yaml"])
+    else:
+        update_inventory_compose(
+            [f"{trap_external_ip},1166,2c,public,,,600,in_profile;equals_profile,,t"]
+        )
+        upgrade_docker_compose()
     time.sleep(120)
 
 
@@ -780,6 +1025,7 @@ def setup_single_regex_and_options_profiles(request):
     options_profile should result in polling IF-MIB.ifDescr
     """
     trap_external_ip = request.config.getoption("trap_external_ip")
+    deployment = request.config.getoption("sc4snmp_deployment")
     profiles = {
         "regex_profile": {
             "frequency": 7,
@@ -804,18 +1050,38 @@ def setup_single_regex_and_options_profiles(request):
             ],
         },
     }
-
-    update_profiles(profiles)
-    update_inventory(
-        [f"{trap_external_ip},1166,2c,public,,,600,regex_profile;options_profile,,"]
-    )
-    upgrade_docker_compose()
+    if deployment == "microk8s":
+        update_profiles_microk8s(profiles)
+        update_file_microk8s(
+            [
+                f"{trap_external_ip},1166,2c,public,,,600,regex_profile;options_profile,,",
+            ],
+            "inventory.yaml",
+        )
+        upgrade_helm_microk8s(["inventory.yaml", "profiles.yaml"])
+    else:
+        update_profiles_compose(profiles)
+        update_inventory_compose(
+            [f"{trap_external_ip},1166,2c,public,,,600,regex_profile;options_profile,,"]
+        )
+        upgrade_docker_compose()
     time.sleep(120)
     yield
-    update_inventory(
-        [f"{trap_external_ip},1166,2c,public,,,600,regex_profile;options_profile,,t"]
-    )
-    upgrade_docker_compose()
+    if deployment == "microk8s":
+        update_file_microk8s(
+            [
+                f"{trap_external_ip},1166,2c,public,,,600,regex_profile;options_profile,,t",
+            ],
+            "inventory.yaml",
+        )
+        upgrade_helm_microk8s(["inventory.yaml"])
+    else:
+        update_inventory_compose(
+            [
+                f"{trap_external_ip},1166,2c,public,,,600,regex_profile;options_profile,,t"
+            ]
+        )
+        upgrade_docker_compose()
     time.sleep(120)
 
 
@@ -855,6 +1121,7 @@ def setup_single_gt_and_lt_profiles_with_negation(request):
     not_lt_profile should result in polling IF-MIB.ifOutDiscards.2
     """
     trap_external_ip = request.config.getoption("trap_external_ip")
+    deployment = request.config.getoption("sc4snmp_deployment")
     profiles = {
         "not_gt_profile": {
             "frequency": 7,
@@ -882,17 +1149,38 @@ def setup_single_gt_and_lt_profiles_with_negation(request):
         },
     }
 
-    update_profiles(profiles)
-    update_inventory(
-        [f"{trap_external_ip},1166,2c,public,,,600,not_gt_profile;not_lt_profile,,"]
-    )
-    upgrade_docker_compose()
+    if deployment == "microk8s":
+        update_profiles_microk8s(profiles)
+        update_file_microk8s(
+            [
+                f"{trap_external_ip},1166,2c,public,,,600,not_gt_profile;not_lt_profile,,",
+            ],
+            "inventory.yaml",
+        )
+        upgrade_helm_microk8s(["inventory.yaml", "profiles.yaml"])
+    else:
+        update_profiles_compose(profiles)
+        update_inventory_compose(
+            [f"{trap_external_ip},1166,2c,public,,,600,not_gt_profile;not_lt_profile,,"]
+        )
+        upgrade_docker_compose()
     time.sleep(120)
     yield
-    update_inventory(
-        [f"{trap_external_ip},1166,2c,public,,,600,not_gt_profile;not_lt_profile,,t"]
-    )
-    upgrade_docker_compose()
+    if deployment == "microk8s":
+        update_file_microk8s(
+            [
+                f"{trap_external_ip},1166,2c,public,,,600,not_gt_profile;not_lt_profile,,t",
+            ],
+            "inventory.yaml",
+        )
+        upgrade_helm_microk8s(["inventory.yaml"])
+    else:
+        update_inventory_compose(
+            [
+                f"{trap_external_ip},1166,2c,public,,,600,not_gt_profile;not_lt_profile,,t"
+            ]
+        )
+        upgrade_docker_compose()
     time.sleep(120)
 
 
@@ -932,6 +1220,7 @@ def setup_single_in_and_equals_profiles_with_negation(request):
     not_equals_profile should result in polling IF-MIB.ifOutDiscards.1
     """
     trap_external_ip = request.config.getoption("trap_external_ip")
+    deployment = request.config.getoption("sc4snmp_deployment")
     profiles = {
         "not_in_profile": {
             "frequency": 7,
@@ -959,19 +1248,40 @@ def setup_single_in_and_equals_profiles_with_negation(request):
         },
     }
 
-    update_profiles(profiles)
-    update_inventory(
-        [f"{trap_external_ip},1166,2c,public,,,600,not_in_profile;not_equals_profile,,"]
-    )
-    upgrade_docker_compose()
+    if deployment == "microk8s":
+        update_profiles_microk8s(profiles)
+        update_file_microk8s(
+            [
+                f"{trap_external_ip},1166,2c,public,,,600,not_in_profile;not_equals_profile,,",
+            ],
+            "inventory.yaml",
+        )
+        upgrade_helm_microk8s(["inventory.yaml", "profiles.yaml"])
+    else:
+        update_profiles_compose(profiles)
+        update_inventory_compose(
+            [
+                f"{trap_external_ip},1166,2c,public,,,600,not_in_profile;not_equals_profile,,"
+            ]
+        )
+        upgrade_docker_compose()
     time.sleep(120)
     yield
-    update_inventory(
-        [
-            f"{trap_external_ip},1166,2c,public,,,600,not_in_profile;not_equals_profile,,t"
-        ]
-    )
-    upgrade_docker_compose()
+    if deployment == "microk8s":
+        update_file_microk8s(
+            [
+                f"{trap_external_ip},1166,2c,public,,,600,not_in_profile;not_equals_profile,,t",
+            ],
+            "inventory.yaml",
+        )
+        upgrade_helm_microk8s(["inventory.yaml"])
+    else:
+        update_inventory_compose(
+            [
+                f"{trap_external_ip},1166,2c,public,,,600,not_in_profile;not_equals_profile,,t"
+            ]
+        )
+        upgrade_docker_compose()
     time.sleep(120)
 
 
@@ -1011,6 +1321,7 @@ def setup_single_regex_and_options_profiles_with_negation(request):
     not_options_profile should result in polling IF-MIB.ifDescr
     """
     trap_external_ip = request.config.getoption("trap_external_ip")
+    deployment = request.config.getoption("sc4snmp_deployment")
     profiles = {
         "not_regex_profile": {
             "frequency": 7,
@@ -1038,21 +1349,40 @@ def setup_single_regex_and_options_profiles_with_negation(request):
         },
     }
 
-    update_profiles(profiles)
-    update_inventory(
-        [
-            f"{trap_external_ip},1166,2c,public,,,600,not_regex_profile;not_options_profile,,"
-        ]
-    )
-    upgrade_docker_compose()
+    if deployment == "microk8s":
+        update_profiles_microk8s(profiles)
+        update_file_microk8s(
+            [
+                f"{trap_external_ip},1166,2c,public,,,600,not_regex_profile;not_options_profile,,",
+            ],
+            "inventory.yaml",
+        )
+        upgrade_helm_microk8s(["inventory.yaml", "profiles.yaml"])
+    else:
+        update_profiles_compose(profiles)
+        update_inventory_compose(
+            [
+                f"{trap_external_ip},1166,2c,public,,,600,not_regex_profile;not_options_profile,,"
+            ]
+        )
+        upgrade_docker_compose()
     time.sleep(120)
     yield
-    update_inventory(
-        [
-            f"{trap_external_ip},1166,2c,public,,,600,not_regex_profile;not_options_profile,,t"
-        ]
-    )
-    upgrade_docker_compose()
+    if deployment == "microk8s":
+        update_file_microk8s(
+            [
+                f"{trap_external_ip},1166,2c,public,,,600,not_regex_profile;not_options_profile,,t",
+            ],
+            "inventory.yaml",
+        )
+        upgrade_helm_microk8s(["inventory.yaml"])
+    else:
+        update_inventory_compose(
+            [
+                f"{trap_external_ip},1166,2c,public,,,600,not_regex_profile;not_options_profile,,t"
+            ]
+        )
+        upgrade_docker_compose()
     time.sleep(120)
 
 
@@ -1096,6 +1426,7 @@ def setup_multiple_conditions_profiles(request):
     lt_and_in_profile should result in polling IF-MIB.ifOutDiscards.2
     """
     trap_external_ip = request.config.getoption("trap_external_ip")
+    deployment = request.config.getoption("sc4snmp_deployment")
     profiles = {
         "gt_and_equals_profile": {
             "frequency": 7,
@@ -1123,21 +1454,40 @@ def setup_multiple_conditions_profiles(request):
         },
     }
 
-    update_profiles(profiles)
-    update_inventory(
-        [
-            f"{trap_external_ip},1166,2c,public,,,600,gt_and_equals_profile;lt_and_in_profile,,"
-        ]
-    )
-    upgrade_docker_compose()
+    if deployment == "microk8s":
+        update_profiles_microk8s(profiles)
+        update_file_microk8s(
+            [
+                f"{trap_external_ip},1166,2c,public,,,600,gt_and_equals_profile;lt_and_in_profile,,",
+            ],
+            "inventory.yaml",
+        )
+        upgrade_helm_microk8s(["inventory.yaml", "profiles.yaml"])
+    else:
+        update_profiles_compose(profiles)
+        update_inventory_compose(
+            [
+                f"{trap_external_ip},1166,2c,public,,,600,gt_and_equals_profile;lt_and_in_profile,,"
+            ]
+        )
+        upgrade_docker_compose()
     time.sleep(120)
     yield
-    update_inventory(
-        [
-            f"{trap_external_ip},1166,2c,public,,,600,gt_and_equals_profile;lt_and_in_profile,,t"
-        ]
-    )
-    upgrade_docker_compose()
+    if deployment == "microk8s":
+        update_file_microk8s(
+            [
+                f"{trap_external_ip},1166,2c,public,,,600,gt_and_equals_profile;lt_and_in_profile,,t",
+            ],
+            "inventory.yaml",
+        )
+        upgrade_helm_microk8s(["inventory.yaml"])
+    else:
+        update_inventory_compose(
+            [
+                f"{trap_external_ip},1166,2c,public,,,600,gt_and_equals_profile;lt_and_in_profile,,t"
+            ]
+        )
+        upgrade_docker_compose()
     time.sleep(120)
 
 
@@ -1180,6 +1530,7 @@ def setup_wrong_conditions_profiles(request):
     None of the profiles below should poll anything.
     """
     trap_external_ip = request.config.getoption("trap_external_ip")
+    deployment = request.config.getoption("sc4snmp_deployment")
     profiles = {
         "wrong_gt_and_equals_profile": {
             "frequency": 7,
@@ -1214,21 +1565,40 @@ def setup_wrong_conditions_profiles(request):
         },
     }
 
-    update_profiles(profiles)
-    update_inventory(
-        [
-            f"{trap_external_ip},1166,2c,public,,,600,wrong_gt_and_equals_profile;wrong_lt_and_in_profile;wrong_equals_profile,,"
-        ]
-    )
-    upgrade_docker_compose()
+    if deployment == "microk8s":
+        update_profiles_microk8s(profiles)
+        update_file_microk8s(
+            [
+                f"{trap_external_ip},1166,2c,public,,,600,wrong_gt_and_equals_profile;wrong_lt_and_in_profile;wrong_equals_profile,,",
+            ],
+            "inventory.yaml",
+        )
+        upgrade_helm_microk8s(["inventory.yaml", "profiles.yaml"])
+    else:
+        update_profiles_compose(profiles)
+        update_inventory_compose(
+            [
+                f"{trap_external_ip},1166,2c,public,,,600,wrong_gt_and_equals_profile;wrong_lt_and_in_profile;wrong_equals_profile,,"
+            ]
+        )
+        upgrade_docker_compose()
     time.sleep(120)
     yield
-    update_inventory(
-        [
-            f"{trap_external_ip},1166,2c,public,,,600,wrong_gt_and_equals_profile;wrong_lt_and_in_profile;wrong_equals_profile,,t"
-        ]
-    )
-    upgrade_docker_compose()
+    if deployment == "microk8s":
+        update_file_microk8s(
+            [
+                f"{trap_external_ip},1166,2c,public,,,600,wrong_gt_and_equals_profile;wrong_lt_and_in_profile;wrong_equals_profile,,t",
+            ],
+            "inventory.yaml",
+        )
+        upgrade_helm_microk8s(["inventory.yaml"])
+    else:
+        update_inventory_compose(
+            [
+                f"{trap_external_ip},1166,2c,public,,,600,wrong_gt_and_equals_profile;wrong_lt_and_in_profile;wrong_equals_profile,,t"
+            ]
+        )
+        upgrade_docker_compose()
     time.sleep(120)
 
 
@@ -1237,6 +1607,188 @@ class TestWrongConditions:
     def test_wrong_profiles(self, request, setup_splunk):
         time.sleep(20)
         search_string = """| mpreview index=netmetrics | search profiles=wrong_gt_and_equals_profile OR profiles=wrong_lt_and_in_profile OR profiles=wrong_equals_profile """
+        result_count, metric_count = run_retried_single_search(
+            setup_splunk, search_string, 2
+        )
+        assert result_count == 0
+        assert metric_count == 0
+
+
+@pytest.fixture(scope="class")
+def setup_misconfigured_profiles(request):
+    """
+    None of the profiles below should poll anything.
+    """
+    trap_external_ip = request.config.getoption("trap_external_ip")
+    deployment = request.config.getoption("sc4snmp_deployment")
+    profiles = {
+        "no_varbinds_profile": {
+            "frequency": 7,
+            "varBinds": [],
+            "conditions": [
+                {"field": "IF-MIB.ifIndex", "operation": dq("gt"), "value": 20},
+                {
+                    "field": "IF-MIB.ifDescr",
+                    "operation": dq("equals"),
+                    "value": dq("eth0"),
+                },
+            ],
+        },
+        "no_operation_key_in_condition_profile": {
+            "frequency": 7,
+            "varBinds": [yaml_escape_list(sq("IF-MIB"), sq("ifOutDiscards"))],
+            "conditions": [
+                {"field": "IF-MIB.ifIndex", "operation": dq("lt"), "value": 20},
+                {
+                    "field": "IF-MIB.ifDescr",
+                    "value": [dq("test value 1"), dq("test value 2")],
+                },
+            ],
+        },
+        "no_frequency_profile": {
+            "varBinds": [yaml_escape_list(sq("IF-MIB"), sq("ifOutDiscards"))],
+            "conditions": [
+                {"field": "IF-MIB.ifIndex", "operation": dq("equals"), "value": 200}
+            ],
+        },
+        "no_patterns_profile": {
+            "frequency": 3,
+            "condition": {
+                "type": "field",
+                "field": "SNMPv2-MIB.sysDescr",
+            },
+            "varBinds": [
+                yaml_escape_list(sq("IP-MIB"), sq("icmpOutDestUnreachs"), 0),
+                yaml_escape_list(sq("IP-MIB"), sq("icmpOutEchoReps"), 0),
+            ],
+        },
+    }
+
+    if deployment == "microk8s":
+        update_profiles_microk8s(profiles)
+        update_file_microk8s(
+            [
+                f"{trap_external_ip},1165,2c,public,,,600,no_varbinds_profile;no_operation_key_in_condition_profile;no_frequency_profile;no_patterns_profile,,",
+            ],
+            "inventory.yaml",
+        )
+        upgrade_helm_microk8s(["inventory.yaml", "profiles.yaml"])
+    else:
+        update_profiles_compose(profiles)
+        update_inventory_compose(
+            [
+                f"{trap_external_ip},1165,2c,public,,,600,no_varbinds_profile;no_operation_key_in_condition_profile;no_frequency_profile;no_patterns_profile,,",
+            ]
+        )
+        upgrade_docker_compose()
+    time.sleep(120)
+    yield
+    if deployment == "microk8s":
+        update_file_microk8s(
+            [
+                f"{trap_external_ip},1165,2c,public,,,600,no_varbinds_profile;no_operation_key_in_condition_profile;no_frequency_profile;no_patterns_profile,,t",
+            ],
+            "inventory.yaml",
+        )
+        upgrade_helm_microk8s(["inventory.yaml"])
+    else:
+        update_inventory_compose(
+            [
+                f"{trap_external_ip},1165,2c,public,,,600,no_varbinds_profile;no_operation_key_in_condition_profile;no_frequency_profile;no_patterns_profile,,t",
+            ]
+        )
+        upgrade_docker_compose()
+    time.sleep(120)
+
+
+@pytest.mark.usefixtures("setup_misconfigured_profiles")
+class TestMisconfiguredProfiles:
+    def test_wrong_profiles(self, request, setup_splunk):
+        time.sleep(20)
+        search_string = """| mpreview index=netmetrics | search profiles=no_varbinds_profile OR profiles=no_operation_key_in_condition_profile OR profiles=no_frequency_profile OR profiles=no_patterns_profile """
+        result_count, metric_count = run_retried_single_search(
+            setup_splunk, search_string, 2
+        )
+        assert result_count == 0
+        assert metric_count == 0
+
+
+@pytest.fixture(scope="class")
+def setup_misconfigured_groups(request):
+    trap_external_ip = request.config.getoption("trap_external_ip")
+    deployment = request.config.getoption("sc4snmp_deployment")
+    profiles = {
+        "routers_wrong_group_profile": {
+            "frequency": 7,
+            "varBinds": [yaml_escape_list(sq("IP-MIB"))],
+        },
+        "switches_wrong_group_profile": {
+            "frequency": 7,
+            "varBinds": [yaml_escape_list(sq("IP-MIB"))],
+        },
+    }
+    groups = {
+        "routers": [
+            {"address": trap_external_ip, "port": 1163},
+            {"address": trap_external_ip, "port": 1164, "wrong_key": 1},
+        ],
+        "switches": [
+            {"addre": trap_external_ip},
+            {"address": trap_external_ip, "port": 1162},
+        ],
+    }
+    if deployment == "microk8s":
+        update_profiles_microk8s(profiles)
+        update_groups_microk8s(groups)
+        update_file_microk8s(
+            [
+                f"{trap_external_ip},1165,2c,public,,,600,single_profile,,",
+                f"routers,,2c,public,,,600,routers_wrong_group_profile,,",
+                f"switches,,2c,public,,,600,switches_wrong_group_profile,,",
+            ],
+            "inventory.yaml",
+        )
+        upgrade_helm_microk8s(["inventory.yaml", "profiles.yaml", "groups.yaml"])
+    else:
+        update_profiles_compose(profiles)
+        update_groups_compose(groups)
+        update_inventory_compose(
+            [
+                f"{trap_external_ip},1165,2c,public,,,600,single_profile,,",
+                f"routers,,2c,public,,,600,routers_wrong_group_profile,,",
+                f"switches,,2c,public,,,600,switches_wrong_group_profile,,",
+            ]
+        )
+        upgrade_helm_microk8s(["inventory.yaml"])
+    time.sleep(120)
+    yield
+    if deployment == "microk8s":
+        update_file_microk8s(
+            [
+                f"{trap_external_ip},1165,2c,public,,,600,single_profile,,t",
+                f"routers,,2c,public,,,600,routers_wrong_group_profile,,t",
+                f"switches,,2c,public,,,600,switches_wrong_group_profile,,t",
+            ],
+            "inventory.yaml",
+        )
+        upgrade_helm_microk8s(["inventory.yaml"])
+    else:
+        update_inventory_compose(
+            [
+                f"{trap_external_ip},1165,2c,public,,,600,single_profile,,t",
+                f"routers,,2c,public,,,600,routers_wrong_group_profile,,t",
+                f"switches,,2c,public,,,600,switches_wrong_group_profile,,t",
+            ]
+        )
+        upgrade_docker_compose()
+    time.sleep(100)
+
+
+@pytest.mark.usefixtures("setup_misconfigured_groups")
+class TestMisconfiguredGroups:
+    def test_wrong_groups(self, request, setup_splunk):
+        time.sleep(20)
+        search_string = """| mpreview index=netmetrics | search profiles=routers_wrong_group_profile OR profiles=routers_wrong_group_profile """
         result_count, metric_count = run_retried_single_search(
             setup_splunk, search_string, 2
         )
